Refactory and improved block monitoring struct
