+++ config.json
@@ -1,22 +1,28 @@
 {
-  "api": "",
-  "watcher" : "",
-  "validators": [],
-  "tg_group":,
-  "tg_bot_token":"",
-  "slack_hook":"",
-  "slack_users": {
-    "tkivaloper1....": {
-      "slack_username": "<@U...>",
-      "alert_threshold": "warning"
+  "valmonitor": {
+    "api": "",
+    "watcher": "",
+    "test_interval_min": 10,
+    "state_file": "./state.json",
+    "validators": [],
+    "slack_users": {},
+    "alert_thresholds": {
+      "notice": 20,
+      "warning": 200,
+      "error": 300,
+      "critical": 400
     }
   },
-  "alert_thresholds": {
-    "notice": 10,
-    "warning": 20,
-    "error": 30,
-    "critical": 40
+  "blcmonitor": {
+    "apis": {},
+    "watcher": "",
+    "test_interval_min": 5
   },
+
+  "tg_group": ,
+  "tg_bot_token": "",
+  "slack_hook": "",
+
   "emoji": {
     "notice": "0x2139",
     "warning": "0x26A0",
