+++ monitor.js
@@ -1,276 +1,14 @@
-const https = require('https');
-const axios = require('axios');
 const config = require(process.argv[2]);
-const api = config.api;
-const emoji = config.emoji;
-const watcher = config.watcher;
-const tg_group = config.tg_group;
+const run_options = process.argv[3]; // 0: validator 1:block 2:both
+
+
 const tg_bot_token = config.tg_bot_token;
-const slack_hook = config.slack_hook;
-const slack_users = config.slack_users;
-const validators_to_watch = config.validators
-const alert_thresholds = config.alert_thresholds;
-const update_all = true
-const test_interval_min = process.argv[4]
-
-var test_interval_block = test_interval_min * 60 / 5
-var state = require(process.argv[3]);
-
-
-// Get all active validators
-async function getAllActiveValidators() {
-  let validators = []
-  let validator_data = {}
-
-  await axios.get(api + "/staking/validators").then(function(response) {
-      validators = response.data.result
-    })
-    .catch(function(error) {
-      // handle error
-      console.log(error);
-    })
-
-  for (validator of validators) {
-    validator_data[validator.operator_address] = {
-      moniker: validator.description.moniker,
-      pubkey: validator.consensus_pubkey
-    }
-  }
-  return validator_data;
-}
-
-// Get active validator consensus addresses
-async function getAllActiveValidatorConsAddress() {
-  let validators = []
-  let addresses = {}
-  await axios.get(api + "/validatorsets/latest").then(function(response) {
-      validators = response.data.result
-    })
-    .catch(function(error) {
-      // handle error
-      console.log(error);
-    })
-
-  for (validator of validators.validators) {
-    addresses[validator.pub_key] = validator.address
-  }
-
-  return addresses;
-}
-
-// Get missed block infos
-async function getSigningInfo() {
-  let signing_info = []
-  let missed_blocks = {}
-  await axios.get(api + "/slashing/signing_infos").then(function(response) {
-      signing_info = response.data.result
-    })
-    .catch(function(error) {
-      console.log(error);
-    })
-
-  for (validator of signing_info) {
-    missed_blocks[validator.address] = validator.missed_blocks_counter
-  }
-  return missed_blocks;
-}
-
-// Format the alert message and send it to slack
-function sendAlertsSlack(data) {
-  let messageBody = {
-    "text": ""
-  }
-  for (line of data) {
-    temp_tag = line[3] ? line[2] : ""
-    messageBody.text += String.fromCodePoint(emoji[line[4]]) + " *" + line[0] + "* has missed " + line[1] + " blocks over the last 5000 blocks " + temp_tag + "\n"
-  }
-
-  try {
-    messageBody = JSON.stringify(messageBody);
-  } catch (e) {
-    throw new Error('Failed to stringify messageBody', e);
-  }
-
-  // Promisify the https.request
-  return new Promise((resolve, reject) => {
-    // general request options, we defined that it's a POST request and content is JSON
-    const requestOptions = {
-      method: 'POST',
-      header: {
-        'Content-Type': 'application/json'
-      }
-    };
-
-    // actual request
-    const req = https.request(slack_hook, requestOptions, (res) => {
-      let response = '';
-
-
-      res.on('data', (d) => {
-        response += d;
-      });
-
-      // response finished, resolve the promise with data
-      res.on('end', () => {
-        resolve(response);
-      })
-    });
-
-    // there was an error, reject the promise
-    req.on('error', (e) => {
-      reject(e);
-    });
-
-    // send our message body (was parsed to JSON beforehand)
-    req.write(messageBody);
-    req.end();
-  });
-}
-
-// Format the alert message and send it to Telegram
-function sendAlertsTelegram(data) {
-  let messageBody = {
-    "text": ""
-  }
-
-  for (line of data) {
-    temp_tag = line[3] ? line[2] : ""
-    // remove reserved char form moniker:
-    moniker = line[0].replace("_", "\_")
-                      .replace("-", "\-")
-                      .replace(".", "\.")
-                      .replace("*", "\*")
-                      .replace("[", "\[")
-                      .replace("`", "\`");
-    messageBody.text += String.fromCodePoint(emoji[line[4]]) + " <b>" + moniker + "</b> has missed " + line[1] + " blocks over the last 5000 blocks \n -------- \n "
-  }
-
-  telegram.sendMessage(tg_group, messageBody.text, {parse_mode:"HTML"})
-}
-
-// Get the level of severity of the downtime
-function getSeverity(missed) {
-  var sev = 0;
-  var i = 0;
-  var keys = Object.keys(alert_thresholds)
-  while (alert_thresholds[keys[i]] <= missed) {
-    sev = keys[i];
-    i++;
-  }
-  return sev
-}
-
-// ping the watcher to continuously ensure the script is up
-async function pingWatcher() {
-  https.get(watcher).on('error', (err) => {
-    console.log('Ping failed: ' + err)
-  });
-}
-
-// Save the last state of each validator
-function saveState(data) {
-  var fs = require('fs');
-  fs.writeFile(process.argv[3], JSON.stringify(data), 'utf8', function(err){
-    if(err) throw err;
-  });
-
-  return data
-}
-
-// Run one cycle of monitoring
-async function runMonitor() {
-  var now = new Date()
-  console.log("\n" + now.toISOString() + ": Satrting a monitoring cycle \n")
-
-  var validators_to_alert = []
-  var temp_tag = ''
-  var temp_mention = ''
-  var new_state = {}
-
-  if (update_all) {
-    var active_validators = await getAllActiveValidators()
-    var active_validators_cons = await getAllActiveValidatorConsAddress()
-  }
-
-  var signing_info = await getSigningInfo()
-  if (validators_to_watch.length != 0) {
-    var vtw = validators_to_watch
-  } else {
-    var vtw = Object.keys(active_validators)
-  }
-
-  for (val of vtw) {
-    // check the number of missed block
-    var current_missed = {
-                      missed : signing_info[active_validators_cons[active_validators[val].pubkey]],
-                      alerts : 1
-                      }
-
-    // for testing
-    // var current_missed = {missed : Math.floor(Math.random() * 600), alerts : 1 }
-
-    // prepare the new state
-    new_state[val] = current_missed
-
-    // get the old state
-    temp_old_state_missed = (state[val] != undefined) ? state[val].missed : 0 // if first run
-    temp_old_state_alerts = (state[val] != undefined) ? state[val].alerts : 1 // if first run
-
-    // if current missed is greater than the threshold adjusted with the alert offset send an alert on slack
-    if (current_missed.missed >= parseInt(temp_old_state_missed) + Math.min(alert_thresholds['notice'] * temp_old_state_alerts, test_interval_block) ) {
-      console.log( "Added validator to the alert list : " + active_validators[val].moniker)
-
-      // update the alert counter
-      new_state[val].alerts = temp_old_state_alerts + 1
-
-      // Tag if the user chose to recieve an alert for the current number of missed blocks
-      if (Object.keys(slack_users).includes(val)) {
-        temp_tag = (current_missed.missed >= alert_thresholds[slack_users[val].alert_threshold]);
-        temp_mention = slack_users[val].slack_username;
-      } else {
-        temp_tag = false;
-        temp_mention = 0;
-      }
-      // Prepare the message data for the slack alert function
-      validators_to_alert.push([active_validators[val].moniker, current_missed.missed, temp_mention, temp_tag, getSeverity(current_missed.missed)])
-    }
-    else{
-      // if it is less than the old state reset the alert counter
-      if (current_missed.missed < temp_old_state_missed) {
-        new_state[val].alerts = 1
-      }
-      // if it is greater than the old state (unadjusted) maintain the old alert counter
-      else{
-        new_state[val].alerts = temp_old_state_alerts
-      }
-    }
-  }
-
-  if (validators_to_alert.length == 0){
-    console.log( "No validator needs to be alerted")
-  }
-  else{
-    // send the alerts
-    await sendAlertsSlack(validators_to_alert)
-    await sendAlertsTelegram(validators_to_alert)
-  }
-
-
-  // save the state
-  state = saveState(new_state)
-
-  // ping the watcher
-  await pingWatcher();
-
-  now = new Date()
-  console.log("\n" + now.toISOString() + ": Monitoring cycle ended")
-  console.log("\nWaiting for the next cycle ...")
-
-  setTimeout(runMonitor, test_interval_min * 60 * 1000)
-}
+const {runBlockMonitor} = require("./monitor_block")
+const {runValidatorMonitor} = require("./monitor_validator")
 
 var TelegramBot = require("node-telegram-bot-api")
 telegram = new TelegramBot(tg_bot_token, {polling: true});
 
-runMonitor()
 
+if (run_options == 0 || run_options == 2) runValidatorMonitor(config)
+if (run_options == 1 || run_options == 2) runBlockMonitor(config)
